+++ CHANGELOG.md
@@ -49,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * (precompiles) Add slinky precompiled contract
 * (x/warden) Add broadcastType to SignRequest
 * (precompiles) Add broadcastType to SignRequest
-
+* (precompiles) #1077 Change status fields from int32 to enums/uint8 in precompile ABIs for improved type safety and consistency
 
 ### Bug Fixes